// src/generated/attachment.ts
export interface AttachmentRequest {
  id?: string | undefined;
  type?: Attachment_AttachmentType | undefined;
  category: Attachment_AttachmentCategory;
  url: string;
  fileName: string;
  fileUrl: string;
  fileData?: Uint8Array | undefined;
  isVisible: boolean;
  priority: number;
}